modify result

+++ http/rpc/rpc.go
@@ -107,7 +107,7 @@ func Handle(w http.ResponseWriter, r *http.Request) {
 		//if the function does not exist
 		RpcLog.Warn("HTTP JSON RPC Handle - No function to call for ", request["method"])
 		data, err := json.Marshal(map[string]interface{}{
-			"errorCode": -32601,
+			"errorCode": int64(-32601),
 			"desc":      "The called method was not found on the server",
 			"result":    nil,
 		})
